Restart the test suite using child processes
This ensures tests are run in their own environment (no shared console.log, etc.) and that the actual Mocha CLI + compiler is run and not a programmatic API that could do something different, as the programmatic API is little documented.

As a drawback, tests are now really slow, ~200-300ms per test, but I'll worry about this when it's actually a problem.