Add support and tests for usage with Jasmine
Nothing needed changing in the code, but as I was very unfamiliar with Jasmine, it took me quite some time to figure this out.
An almost identical test suite now ensures using fulky against Jasmine works correctly, but the default reporter (and only one that ships by default with Jasmine) is not good enough. I will improve this in the future.

Also, add a tiny bit of documentation in the README file.

BREAKING CHANGE: With Mocha, now use `--compilers=md:fulky/md-compiler`.